[#32] refactor : Change Id type from Integer to Long
- id 타입을 Integer에서 Long으로 변경
- 디폴트 생성자 제거
- 주석처리 된 메소드 제거

+++ backend/src/main/java/com/team15/sidedish/controller/BestController.java
@@ -29,7 +29,7 @@ public class BestController {
     }
 
     @GetMapping("/{categoryId}")
-    public BestDTO showSingleBestDish(@PathVariable Integer categoryId) {
+    public BestDTO showSingleBestDish(@PathVariable Long categoryId) {
         return bestService.showSingleBestDish(categoryId);
     }
 }
